fixed rating, hope it does break

+++ part3/hbnb/app/models/review.py
@@ -12,6 +12,6 @@ class Review(BaseModel):
     __tablename__ = 'reviews'
 
     text = db.Column(db.String(256), nullable=False)
-    rating = db.Column(db.Integer, nullable=False)
+    rating = db.Column(db.Integer(5), nullable=False)
     user_id = db.Column(db.String(36), db.ForeignKey('users.id'), nullable=False)
     place_id = db.Column(db.String(36), db.ForeignKey('places.id'), nullable=False)
